chore: Minor cleanup of Speckle.Automate.Sdk
