+++ Automate/Speckle.Automate.Sdk/Schema/AutomationStatus.cs
@@ -1,0 +1,12 @@
+namespace Speckle.Automate.Sdk.Schema;
+
+///<summary>
+/// Set the status of the automation.
+///</summary>
+public enum AutomationStatus
+{
+  Initializing,
+  Running,
+  Failed,
+  Succeeded
+}
