+++ Automate/Speckle.Automate.Sdk/Schema/BlobUploadResponse.cs
@@ -1,0 +1,6 @@
+namespace Speckle.Automate.Sdk.Schema;
+
+public struct BlobUploadResponse
+{
+  public List<UploadResult> UploadResults { get; set; }
+}
