+++ Automate/Speckle.Automate.Sdk/Schema/ObjectResultLevel.cs
@@ -1,0 +1,8 @@
+namespace Speckle.Automate.Sdk.Schema;
+
+public enum ObjectResultLevel
+{
+  Info,
+  Warning,
+  Error
+}
