+++ Automate/Speckle.Automate.Sdk/Schema/ObjectResultValues.cs
@@ -1,0 +1,7 @@
+namespace Speckle.Automate.Sdk.Schema;
+
+struct ObjectResultValues
+{
+  public List<ResultCase> ObjectResults { get; set; }
+  public List<string> BlobIds { get; set; }
+}
