+++ Automate/Speckle.Automate.Sdk/Schema/ObjectResults.cs
@@ -1,0 +1,7 @@
+namespace Speckle.Automate.Sdk.Schema;
+
+struct ObjectResults
+{
+  public string Version => "1.0.0";
+  public ObjectResultValues Values { get; set; }
+}
