+++ Automate/Speckle.Automate.Sdk/Schema/UploadResult.cs
@@ -1,0 +1,8 @@
+namespace Speckle.Automate.Sdk.Schema;
+
+public struct UploadResult
+{
+  public string BlobId { get; set; }
+  public string FileName { get; set; }
+  public int UploadStatus { get; set; }
+}
